+++ packages/base/src/cli.ts
@@ -3,6 +3,7 @@ import {CommandClass} from 'clipanion'
 import {CloudRunFlareCommand} from './commands/cloud-run/flare'
 import {InstrumentCommand as CloudRunInstrumentCommand} from './commands/cloud-run/instrument'
 import {UninstrumentCommand as CloudRunUninstrumentCommand} from './commands/cloud-run/uninstrument'
+import {DORADeploymentCommand} from './commands/dora/deployment-command'
 import {UploadCommand} from './commands/git-metadata/upload'
 import {LambdaFlareCommand} from './commands/lambda/flare'
 import {InstrumentCommand as LambdaInstrumentCommand} from './commands/lambda/instrument'
@@ -24,4 +25,5 @@ export const baseCommands: Record<string, CommandClass[]> = {
   lambda: [LambdaInstrumentCommand, LambdaUninstrumentCommand, LambdaFlareCommand],
   stepfunctions: [InstrumentStepFunctionsCommand, UninstrumentStepFunctionsCommand],
   'git-metadata': [UploadCommand],
+  dora: [DORADeploymentCommand],
 }
